deprecate non-integer n in nbdtr, nbdtrc, nbdtri
modifies cephes code to use double-valued k as done previously
for bdtr, bdtrc, bdtri in gh-11045

warning message in _legacy_deprecation is updated to specify
removal in SciPy 1.11.

remove unused arguments to _legacy_deprecation

update nbdtr signatures in test_cython_special.py

add test_nbdtr.py to meson.build

+++ scipy/special/cephes.h
@@ -84,9 +84,9 @@ extern double k1(double x);
 extern double k1e(double x);
 extern double kn(int nn, double x);
 
-extern double nbdtrc(int k, int n, double p);
-extern double nbdtr(int k, int n, double p);
-extern double nbdtri(int k, int n, double p);
+extern double nbdtrc(double k, int n, double p);
+extern double nbdtr(double k, int n, double p);
+extern double nbdtri(double k, int n, double p);
 
 extern double ndtr(double a);
 extern double log_ndtr(double a);
